+++ packages/vscode-plugin/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  [[`383feed`](https://github.com/Ripple-TS/ripple/commit/383feed84b09541c0b58992c09816b5a15c2d2d8),
+  [`fcd25aa`](https://github.com/Ripple-TS/ripple/commit/fcd25aa549db0d56ccbd596b657b856a5061e20f),
+  [`30126c7`](https://github.com/Ripple-TS/ripple/commit/30126c753c3a08809bacd07c8cf2eca84e8f8cbb),
+  [`b8cd7c4`](https://github.com/Ripple-TS/ripple/commit/b8cd7c4195505976995033a8e369502996f345ad),
+  [`b8cd7c4`](https://github.com/Ripple-TS/ripple/commit/b8cd7c4195505976995033a8e369502996f345ad),
+  [`fee8620`](https://github.com/Ripple-TS/ripple/commit/fee8620fa4e82a7c7e4adb3e434e9db552a3e157),
+  [`2fcacb4`](https://github.com/Ripple-TS/ripple/commit/2fcacb471d7780074f92b20c9b394f7650a941bb),
+  [`8e2aa8e`](https://github.com/Ripple-TS/ripple/commit/8e2aa8e75678c9ebc9b72055f4da474c82a8e834)]:
+  - @tsrx/typescript-plugin@0.3.34
+  - @tsrx/react@0.1.8
+  - @ripple-ts/language-server@0.3.34
+  - @tsrx/ripple@0.0.16
+
 ## 1.0.9
 
 ### Patch Changes